# Conflicts:
#	.github/workflows/native-unit-test.yml
#	.vscode/settings.json
#	native/CMakeLists.txt
#	native/cmake/predefine.cmake
#	native/cocos/audio/android/AudioDecoderSLES.h
#	native/cocos/audio/android/AudioEngine-inl.cpp
#	native/cocos/audio/android/AudioPlayerProvider.cpp
#	native/cocos/audio/android/AudioResampler.h
#	native/cocos/bindings/auto/jsb_cocos_auto.cpp
#	native/cocos/bindings/auto/jsb_cocos_auto.h
#	native/cocos/bindings/jswrapper/config.h
#	native/cocos/bindings/manual/jsb_conversions.h
#	native/cocos/bindings/manual/jsb_dragonbones_manual.cpp
#	native/cocos/bindings/manual/jsb_global.cpp
#	native/cocos/bindings/manual/jsb_module_register.cpp
#	native/cocos/engine/Engine.cpp
#	native/cocos/platform/interfaces/modules/Device.cpp
#	native/extensions/assets-manager/AssetsManagerEx.cpp
#	native/external-config.json
#	platforms/native/engine/jsb-loader.js
#	platforms/native/engine/jsb-scene.js
#	scripts/build-engine/package-lock.json
#	scripts/build-engine/package.json
#	templates/cmake/openharmony.cmake

+++ native/utils/generate_compile_commands_ios.sh
@@ -8,4 +8,6 @@ cmake -Sbuild -Bbuild -DCC_USE_GLES2=OFF -DCC_USE_VULKAN=OFF -DCC_USE_GLES3=OFF 
     -DCMAKE_OSX_SYSROOT=iphoneos \
     -DCMAKE_SYSTEM_NAME=iOS \
     -GXcode
-cp build/compile_commands.json .
+if [[ -f build/compile_commands.json ]]; then
+   cp build/compile_commands.json .
+fi
